refactor(app): compose built-in auth extensions
