+++ crates/coral-app/Cargo.toml
@@ -26,6 +26,7 @@ tonic-types.workspace = true
 tracing.workspace = true
 
 coral-api.workspace = true
+coral-auth-aws = { path = "../auth/aws" }
 coral-spec.workspace = true
 coral-engine.workspace = true
