+++ crates/coral-cli/src/bootstrap.rs
@@ -1,12 +1,6 @@
-use std::sync::Arc;
-
-use coral_auth_aws::AwsSigV4Authenticator;
 use coral_client::{
     AppClient, ClientError,
-    local::{
-        EngineExtensions, EngineExtensionsProvider, LocalServerError, QuerySource, RunningServer,
-        ServerBuilder,
-    },
+    local::{LocalServerError, RunningServer, ServerBuilder},
 };
 
 pub(crate) struct Bootstrap {
@@ -30,28 +24,12 @@ pub(crate) async fn bootstrap() -> Result<Bootstrap, BootstrapError> {
         });
     }
 
-    let server = ServerBuilder::new()
-        .with_engine_extensions_provider(Arc::new(AwsEngineExtensionsProvider))
-        .start()
-        .await?;
+    let server = ServerBuilder::new().start().await?;
     let app = AppClient::connect(server.endpoint_uri()).await?;
     Ok(Bootstrap {
         app,
         _server: Some(server),
     })
-}
-
-#[derive(Debug)]
-struct AwsEngineExtensionsProvider;
-
-impl EngineExtensionsProvider for AwsEngineExtensionsProvider {
-    fn extensions_for(&self, _selected_sources: &[QuerySource]) -> EngineExtensions {
-        let mut extensions = EngineExtensions::default();
-        extensions
-            .request_authenticators
-            .insert("aws_sigv4".to_string(), Arc::new(AwsSigV4Authenticator));
-        extensions
-    }
 }
 
 #[cfg(feature = "cli-test-server")]
@@ -62,20 +40,4 @@ fn bootstrap_endpoint() -> Option<String> {
 #[cfg(not(feature = "cli-test-server"))]
 fn bootstrap_endpoint() -> Option<String> {
     None
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn cli_provider_registers_aws_sigv4() {
-        let extensions = AwsEngineExtensionsProvider.extensions_for(&[]);
-        let authenticator = extensions
-            .request_authenticators
-            .get("aws_sigv4")
-            .expect("cli should register aws authenticator");
-
-        assert_eq!(authenticator.name(), "aws_sigv4");
-    }
 }
